+++ shape/component.go
@@ -43,7 +43,7 @@ func (r *Component) Direction() Direction { return LR }
 func (r *Component) SetClass(c string)    { r.class = c }
 
 func (r *Component) WriteSvg(out io.Writer) error {
-	w, err := draw.NewTagPrinter(out)
+	w, err := draw.NewTagWriter(out)
 	w.Printf(
 		`<rect class="%s" x="%v" y="%v" width="%v" height="%v"/>`,
 		r.class, r.X, r.Y, r.Width(), r.Height())