+++ shape/dot.go
@@ -39,7 +39,7 @@ func (c *Dot) Direction() Direction  { return LR }
 func (c *Dot) SetClass(class string) { c.class = class }
 
 func (c *Dot) WriteSvg(out io.Writer) error {
-	w, err := draw.NewTagPrinter(out)
+	w, err := draw.NewTagWriter(out)
 	x, y := c.Position()
 	x += c.Radius
 	y += c.Radius