+++ shape/state.go
@@ -37,7 +37,7 @@ func (r *State) Direction() Direction { return LR }
 func (r *State) SetClass(c string)    { r.class = c }
 
 func (r *State) WriteSvg(out io.Writer) error {
-	w, err := draw.NewTagPrinter(out)
+	w, err := draw.NewTagWriter(out)
 	w.Printf(
 		`<rect class="%s" x="%v" y="%v" width="%v" height="%v"/>`,
 		r.class, r.X, r.Y, r.Width(), r.Height())