+++ shape/triangle.go
@@ -31,7 +31,7 @@ func (t *Triangle) Direction() Direction { return LR }
 func (t *Triangle) SetClass(c string)    { t.class = c }
 
 func (t *Triangle) WriteSvg(out io.Writer) error {
-	w, err := draw.NewTagPrinter(out)
+	w, err := draw.NewTagWriter(out)
 	// the path is drawn as if it points straight to the right
 	w.Printf(`<path class="%s" d="M%v,%v l-8,-4 l 0,8 Z" />`,
 		t.class, t.x, t.y)